+++ NEWS.md
@@ -1,3 +1,4 @@
+- added `na_distinct = FALSE` argument to `check_key`
 - replace internal `deparse_x_name()` with exported `chk_deparse()`
 - added `check_name()` to check if elements of character vector are each a syntactically valid name
 - `check_named()` now only gives 1 warning if error = FALSE and not named